Prefer artifacts with correct type to transformations
In the FauxAndroid test, this change demonstrates that a library project
can optionally expose a 'classpath' artifact directly, and this will
take precedence over any dependency transformations.

If a java library doesn't expose a 'classpath' artifact, the default
'jar' artifact will be used (via an IdentityTransform). If a java
library explicitly exposes the classes directory as a 'classpath'
artifact, this will be used instead.