Simplify performance test measurement to the absolute minimum
This is an attempt to find out how much of the flakiness is due to
differences in the environment and how much is actually caused by
the measurement code itself. It'll help us understand better what
we can do about flaky performance tests.

This is not meant to be merged into master.